DjangoBackend/requirements.txt

Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
PyJWT==2.10.1
python-decouple==3.8
huey==2.6.0
psycopg2==2.9.9
redis==5.0.8
channels==4.1.0
daphne==4.1.2
channels-redis==4.2.0
paho-mqtt==2.1.0
